Menambah class baru dan melakukan refactoring
Menambah class baru dan melakukan refactoring untuk class employee dan family
