+++ Person.java
@@ -6,24 +6,11 @@ import java.util.List;
 
 public class Person {
     public String idNumber;
-    private String firstName;
-    private String lastName;
+    public String firstName;
+    public String lastName;
     public String address;
-    public List<String> childNames;
-    public List<String> childIdNumbers;
-    private String spouseName;
-    private String spouseIdNumber;
-        
+    	public boolean isForeigner;
     
-    public void addChild(String childName, String childIdNumber) {
-		childNames.add(childName);
-		childIdNumbers.add(childIdNumber);
-    }
-    
-    public void setSpouse(String spouseName, String spouseIdNumber) {
-		this.spouseName = spouseName;
-		this.spouseIdNumber = idNumber;
-    }
     public enum JenisGender {
             Laki_laki,
             Perempuan
@@ -35,8 +22,7 @@ public class Person {
 		this.lastName = lastName;
                 this.address = address;
                 this.gender = gender;
-                childNames = new LinkedList<String>();
-                childIdNumbers = new LinkedList<String>();
+                
 }
 }
